seperated indent class
