+++ trelby/indent.py
@@ -1,0 +1,22 @@
+# information about one indent level in imported text files.
+class Indent:
+    def __init__(self, indent):
+
+        # indent level, i.e. spaces at the beginning
+        self.indent = indent
+
+        # lines with this indent, leading spaces removed
+        self.lines = []
+
+        # assigned line type, or -1 if not assigned yet.
+        self.lt = -1
+
+        # how many of the lines start with "EXT." or "INT."
+        self.sceneStart = 0
+
+        # how many of the lines have "CUT TO:" or "DISSOLVE TO:"
+        self.trans = 0
+
+        # how many of the lines have a form of "^ +\(.*)$", i.e. are most
+        # likely parentheticals
+        self.paren = 0
