+++ trelby/myimport.py
@@ -13,6 +13,8 @@ import trelby.screenplay as screenplay
 import trelby.titles as titles
 import trelby.util as util
 
+from trelby.indent import Indent
+
 # special linetype that means that indent contains action and scene lines,
 # and scene lines are the ones that begin with "EXT." or "INT."
 SCENE_ACTION = -2
@@ -945,31 +947,6 @@ def findIndent(indDict, func, *vars):
             return v.indent
 
     return -1
-
-
-# information about one indent level in imported text files.
-class Indent:
-    def __init__(self, indent):
-
-        # indent level, i.e. spaces at the beginning
-        self.indent = indent
-
-        # lines with this indent, leading spaces removed
-        self.lines = []
-
-        # assigned line type, or -1 if not assigned yet.
-        self.lt = -1
-
-        # how many of the lines start with "EXT." or "INT."
-        self.sceneStart = 0
-
-        # how many of the lines have "CUT TO:" or "DISSOLVE TO:"
-        self.trans = 0
-
-        # how many of the lines have a form of "^ +\(.*)$", i.e. are most
-        # likely parentheticals
-        self.paren = 0
-
 
 class ImportDlg(wx.Dialog):
     def __init__(self, parent, indents):
